new spec generation
