--- kilosort3/generate_spec.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-SPEC_OUTPUT_FILE="spec.json" APP_IMAGE="magland/pc-kilosort3" APP_EXECUTABLE="/app/main" ./main
