--- spike_sorting_utils/generate_spec.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-SPEC_OUTPUT_FILE="spec.json" APP_IMAGE="magland/pc-spike-sorting-utils" APP_EXECUTABLE="/app/main" ./main
